organize and integrate
