+++ ws_system/utils/parser.py
@@ -1,0 +1,24 @@
+'''
+Choose one operation from below :
+    1 : Get status for the past 7 days
+    2 : Print user summary 
+    3 : Predict user next session duration
+    4 : Fetch new data and update users data and ML model
+    5 : Get top 5 users based on time spent gaming
+    6 : Exit program
+'''
+
+
+def base_case():
+    '''
+    base case of the console with choices of actions
+    '''
+
+
+def print_user_summary_parser():
+    user_id = input('Enter user id:\n')
+    p = input('\nEnter period (yy/mm/dd - yy/mm/dd) :\n')
+    #change that to work with class
+
+    return user_id
+
